Re-define BaseProps and FocusableProps
