+++ packages/bricks/scripts/build.js
@@ -16,7 +16,7 @@ const isDev = process.env.NODE_ENV === "development";
 
 const entryPoints = await fg("src/**/*.{ts,tsx}", {
 	onlyFiles: true,
-	ignore: ["**/*.d.ts", "**/~utils.props.ts"],
+	ignore: ["**/*.d.ts"],
 });
 
 await esbuild.build({
