+++ packages/internal-utils/src/index.ts
@@ -6,4 +6,5 @@
 export * from "./common.js";
 export * from "./dom.js";
 export * from "./hooks.js";
+export * from "./props.js";
 export * from "./react.js";
