+++ packages/mui/scripts/build.js
@@ -19,7 +19,6 @@ const entryPoints = await fg("src/**/*.{ts,tsx}", {
 	ignore: [
 		"**/*.d.ts",
 		"**/types.ts", // This will be used for generating a `types.d.ts` file only.
-		"**/~props.ts",
 	],
 });
